# openapi.json
{
  "components": {
    "schemas": {
      "APIResponse": {
        "description": "Top-level envelope returned by every endpoint (untyped ``data``).",
        "properties": {
          "data": {
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "APIResponse",
        "type": "object"
      },
...
      "AdminLoginRequest": {
        "description": "Admin email + password login (admins have no Instagram identity).",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "AdminLoginRequest",
        "type": "object"
      },
...
      "DataResponse_OkResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OkResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[OkResponse]",
        "type": "object"
      },
...
      "ForgotPasswordRequest": {
        "description": "Enumerate-safe password-reset request (brand or admin).",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Buzz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/admin/forgot-password": {
      "post": {
        "description": "Enumerate-safe admin password-reset request.",
        "operationId": "admin_forgot_password_api_auth_admin_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Admin Forgot Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "description": "Admin email + password login \u2014 the only admin session entry point.\n\nAdmins have neither an Instagram identity nor an invite flow, so without\nthis they are unreachable outside local dev (``dev-login`` 404s off-dev).\nRate-limited per-IP and per-account like the brand path.",
        "operationId": "admin_login_api_auth_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TokenResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Admin Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/admin/reset-password": {
      "post": {
        "description": "Consume an admin password-reset token and set a new password.",
        "operationId": "admin_reset_password_api_auth_admin_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Admin Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/brand/forgot-password": {
      "post": {
        "description": "Enumerate-safe brand password-reset request.",
        "operationId": "brand_forgot_password_api_auth_brand_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Brand Forgot Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/brand/login": {
      "post": {
        "description": "Brand email + password login. Issues JWT tokens on success.\n\nRate-limited per-IP (decorator) and per-account (below) \u2014 the per-account cap\nruns *before* the bcrypt verify so credential-stuffing across rotating IPs\ncan't burn CPU. The cap is generous so a third party can't easily lock out a\nbrand by hammering its email.",
        "operationId": "brand_login_api_auth_brand_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TokenResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Brand Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/brand/reset-password": {
      "post": {
        "description": "Consume a brand password-reset token and set a new password.",
        "operationId": "brand_reset_password_api_auth_brand_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Brand Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/brand/set-password": {
      "post": {
        "description": "Consume a brand invite token, set the password, and start a session.\n\nReturns the same ``TokenResponse`` as login so the SPA lands the brand in\ntheir portal without a separate login step.",
        "operationId": "brand_set_password_api_auth_brand_set_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandSetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TokenResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Brand Set Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/dev-login": {
      "post": {
        "description": "Dev-only shortcut to mint a real session without the Instagram flow.\n\nLocal dev has no Meta credentials and the SPA has no login UI yet (Stage 6),\nso the Stage 4 vertical slice needs a way to obtain a token. This issues the\nsame token pair + refresh cookie as the real flow, but only when\n``ENVIRONMENT == \"development\"`` \u2014 outside dev it 404s (invisible).",
        "operationId": "dev_login_api_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TokenResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Dev Login",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}